+++ src/main/kotlin/org/pathcheck/intellij/cql/CqlPSIFileRoot.kt
@@ -3,9 +3,14 @@ package org.pathcheck.intellij.cql
 import com.intellij.extapi.psi.PsiFileBase
 import com.intellij.openapi.fileTypes.FileType
 import com.intellij.psi.FileViewProvider
+import com.intellij.psi.PsiElement
+import com.intellij.psi.PsiNamedElement
+import org.antlr.intellij.adaptor.SymtabUtils
+import org.antlr.intellij.adaptor.psi.ScopeNode
 import javax.swing.Icon
 
-class CqlPSIFileRoot(viewProvider: FileViewProvider) : PsiFileBase(viewProvider, CqlLanguage) {
+
+class CqlPSIFileRoot(viewProvider: FileViewProvider) : PsiFileBase(viewProvider, CqlLanguage), ScopeNode {
     override fun getFileType(): FileType {
         return CqlFileType
     }
@@ -16,5 +21,31 @@ class CqlPSIFileRoot(viewProvider: FileViewProvider) : PsiFileBase(viewProvider,
 
     override fun getIcon(flags: Int): Icon {
         return CqlIcons.FILE
+    }
+
+    override fun getContext(): ScopeNode? {
+        // Maximum Scope
+        return null
+    }
+
+    /**
+     * Looks for identifiers that match the declaration of the element.
+     */
+    override fun resolve(element: PsiNamedElement): PsiElement? {
+        return listOf(
+            "/library/statement/functionDefinition/identifierOrFunctionIdentifier",
+            "/library/statement/expressionDefinition/identifier",
+            "/library/definition/usingDefinition/qualifiedIdentifier",
+            "/library/definition/usingDefinition/localIdentifier/identifier",
+            "/library/definition/includeDefinition/qualifiedIdentifier",
+            "/library/definition/includeDefinition/localIdentifier/identifier",
+            "/library/definition/codesystemDefinition/identifier",
+            "/library/definition/valuesetDefinition/identifier",
+            "/library/definition/codeDefinition/identifier",
+            "/library/definition/conceptDefinition/identifier",
+            "/library/definition/parameterDefinition/identifier"
+        ).map {
+            SymtabUtils.resolve(this, CqlLanguage, element, it)
+        }.filterNotNull().firstOrNull()
     }
 }